integration-test: add storage tests
When AZURE_STORAGE_ACCOUNT and AZURE_STORAGE_ACCESS_KEY passed,
we now have a scenario that runs by uploading two 512k files to
Azure Blob Storage and configures extension to download it.

We verify by checking test container fs diff to see files are
downloaded and we verify checksum on one of the files.